rename Database.kt to UserDatabaseProvider.kt
