+++ logic/src/commonMain/kotlin/com/wire/kalium/logic/AuthenticatedDataSourceSet.kt
@@ -4,7 +4,7 @@ import com.wire.kalium.cryptography.ProteusClient
 import com.wire.kalium.logic.sync.SyncManager
 import com.wire.kalium.logic.sync.WorkScheduler
 import com.wire.kalium.network.AuthenticatedNetworkContainer
-import com.wire.kalium.persistence.db.Database
+import com.wire.kalium.persistence.db.UserDatabaseProvider
 import com.wire.kalium.persistence.kmm_settings.EncryptedSettingsHolder
 import com.wire.kalium.persistence.kmm_settings.KaliumPreferencesSettings
 
@@ -14,7 +14,7 @@ class AuthenticatedDataSourceSet(
     val proteusClient: ProteusClient,
     val workScheduler: WorkScheduler,
     val syncManager: SyncManager,
-    val database: Database,
+    val userDatabaseProvider: UserDatabaseProvider,
     val kaliumPreferencesSettings: KaliumPreferencesSettings,
     val encryptedSettingsHolder: EncryptedSettingsHolder
 )
