+++ logic/src/jvmMain/kotlin/com/wire/kalium/logic/CoreLogic.kt
@@ -6,13 +6,13 @@ import com.wire.kalium.logic.data.session.SessionDataSource
 import com.wire.kalium.logic.data.session.SessionRepository
 import com.wire.kalium.logic.data.user.UserId
 import com.wire.kalium.logic.feature.UserSessionScope
-import com.wire.kalium.logic.network.SessionManagerImpl
 import com.wire.kalium.logic.feature.call.GlobalCallManager
+import com.wire.kalium.logic.network.SessionManagerImpl
 import com.wire.kalium.logic.sync.SyncManagerImpl
 import com.wire.kalium.logic.sync.WorkScheduler
 import com.wire.kalium.network.AuthenticatedNetworkContainer
 import com.wire.kalium.persistence.client.SessionStorageImpl
-import com.wire.kalium.persistence.db.Database
+import com.wire.kalium.persistence.db.UserDatabaseProvider
 import com.wire.kalium.persistence.kmm_settings.EncryptedSettingsHolder
 import com.wire.kalium.persistence.kmm_settings.KaliumPreferencesSettings
 import com.wire.kalium.persistence.kmm_settings.SettingOptions
@@ -38,7 +38,7 @@ actual class CoreLogic(clientLabel: String, rootProteusDirectoryPath: String) : 
             val syncManager = SyncManagerImpl(workScheduler)
             val encryptedSettingsHolder = EncryptedSettingsHolder(SettingOptions.UserSettings(idMapper.toDaoModel(userId)))
             val userPreferencesSettings = KaliumPreferencesSettings(encryptedSettingsHolder.encryptedSettings)
-            val database = Database()
+            val userDatabaseProvider = UserDatabaseProvider()
 
             AuthenticatedDataSourceSet(
                 rootProteusDirectoryPath,
@@ -46,7 +46,7 @@ actual class CoreLogic(clientLabel: String, rootProteusDirectoryPath: String) : 
                 proteusClient,
                 workScheduler,
                 syncManager,
-                database,
+                userDatabaseProvider,
                 userPreferencesSettings,
                 encryptedSettingsHolder
             ).also {
