+++ persistence/src/androidMain/kotlin/com/wire/kalium/persistence/db/UserDatabaseProvider.kt
@@ -1,9 +1,12 @@
 package com.wire.kalium.persistence.db
 
+import android.content.Context
+import android.os.Build
+import android.util.Base64
+import androidx.sqlite.db.SupportSQLiteDatabase
 import app.cash.sqldelight.EnumColumnAdapter
 import app.cash.sqldelight.adapter.primitive.IntColumnAdapter
-import app.cash.sqldelight.db.SqlDriver
-import app.cash.sqldelight.driver.jdbc.sqlite.JdbcSqliteDriver
+import app.cash.sqldelight.driver.android.AndroidSqliteDriver
 import com.wire.kalium.persistence.Client
 import com.wire.kalium.persistence.Conversation
 import com.wire.kalium.persistence.Member
@@ -20,20 +23,41 @@ import com.wire.kalium.persistence.dao.TeamDAO
 import com.wire.kalium.persistence.dao.TeamDAOImpl
 import com.wire.kalium.persistence.dao.UserDAO
 import com.wire.kalium.persistence.dao.UserDAOImpl
+import com.wire.kalium.persistence.dao.UserIDEntity
 import com.wire.kalium.persistence.dao.asset.AssetDAO
 import com.wire.kalium.persistence.dao.asset.AssetDAOImpl
 import com.wire.kalium.persistence.dao.client.ClientDAO
 import com.wire.kalium.persistence.dao.client.ClientDAOImpl
 import com.wire.kalium.persistence.dao.message.MessageDAO
 import com.wire.kalium.persistence.dao.message.MessageDAOImpl
+import com.wire.kalium.persistence.kmm_settings.KaliumPreferences
+import com.wire.kalium.persistence.util.FileNameUtil
+import net.sqlcipher.database.SupportFactory
+import java.security.SecureRandom
 
-actual class Database {
-
-    val database: UserDatabase
+actual class UserDatabaseProvider(private val context: Context, userId: UserIDEntity, kaliumPreferences: KaliumPreferences) {
+    private val dbName = FileNameUtil.userDBName(userId)
+    private val driver: AndroidSqliteDriver
+    private val database: UserDatabase
 
     init {
-        val driver: SqlDriver = JdbcSqliteDriver(JdbcSqliteDriver.IN_MEMORY)
-        UserDatabase.Schema.create(driver)
+        val supportFactory = SupportFactory(getOrGenerateSecretKey(kaliumPreferences).toByteArray())
+
+        val onConnectCallback = object : AndroidSqliteDriver.Callback(UserDatabase.Schema) {
+            override fun onOpen(db: SupportSQLiteDatabase) {
+                super.onOpen(db)
+                db.execSQL("PRAGMA foreign_keys=ON;")
+            }
+        }
+
+        driver = AndroidSqliteDriver(
+            schema = UserDatabase.Schema,
+            context = context,
+            name = dbName,
+            factory = supportFactory,
+            callback = onConnectCallback
+        )
+
         database = UserDatabase(
             driver,
             Client.Adapter(user_idAdapter = QualifiedIDAdapter()),
@@ -51,7 +75,6 @@ actual class Database {
             ),
             User.Adapter(qualified_idAdapter = QualifiedIDAdapter(), IntColumnAdapter)
         )
-        driver.execute(null, "PRAGMA foreign_keys=ON", 0)
     }
 
     actual val userDAO: UserDAO
@@ -76,6 +99,39 @@ actual class Database {
         get() = TeamDAOImpl(database.teamsQueries)
 
     actual fun nuke(): Boolean {
-        TODO("Not yet implemented")
+        driver.close()
+        return context.deleteDatabase(dbName)
     }
+
+    private fun getOrGenerateSecretKey(kaliumPreferences: KaliumPreferences): String {
+        val databaseKey = kaliumPreferences.getString(DATABASE_SECRET_KEY)
+
+        return if (databaseKey == null) {
+            val secretKey = generateSecretKey()
+            kaliumPreferences.putString(DATABASE_SECRET_KEY, secretKey)
+            secretKey
+        } else {
+            databaseKey
+        }
+    }
+
+    private fun generateSecretKey(): String {
+        // TODO review with security
+
+        val random = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            SecureRandom.getInstanceStrong()
+        } else {
+            SecureRandom()
+        }
+        val password = ByteArray(DATABASE_SECRET_LENGTH)
+        random.nextBytes(password)
+
+        return Base64.encodeToString(password, Base64.DEFAULT)
+    }
+
+    companion object {
+        private const val DATABASE_SECRET_KEY = "databaseSecret"
+        private const val DATABASE_SECRET_LENGTH = 48
+    }
+
 }
