+++ persistence/src/commonMain/kotlin/com/wire/kalium/persistence/db/UserDatabaseProvider.kt
@@ -1,14 +1,14 @@
 package com.wire.kalium.persistence.db
 
-import com.wire.kalium.persistence.dao.asset.AssetDAO
 import com.wire.kalium.persistence.dao.ConversationDAO
-import com.wire.kalium.persistence.dao.UserDAO
-import com.wire.kalium.persistence.dao.client.ClientDAO
 import com.wire.kalium.persistence.dao.MetadataDAO
 import com.wire.kalium.persistence.dao.TeamDAO
+import com.wire.kalium.persistence.dao.UserDAO
+import com.wire.kalium.persistence.dao.asset.AssetDAO
+import com.wire.kalium.persistence.dao.client.ClientDAO
 import com.wire.kalium.persistence.dao.message.MessageDAO
 
-expect class Database {
+expect class UserDatabaseProvider {
     val userDAO: UserDAO
     val conversationDAO: ConversationDAO
     val metadataDAO: MetadataDAO
