+++ persistence/src/commonTest/kotlin/com/wire/kalium/persistence/BaseDatabaseTest.kt
@@ -1,10 +1,10 @@
 package com.wire.kalium.persistence
 
-import com.wire.kalium.persistence.db.Database
+import com.wire.kalium.persistence.db.UserDatabaseProvider
 
 expect open class BaseDatabaseTest() {
 
     fun deleteDatabase()
-    fun createDatabase(): Database
+    fun createDatabase(): UserDatabaseProvider
 
 }
