+++ persistence/src/commonTest/kotlin/com/wire/kalium/persistence/dao/UserDAOTest.kt
@@ -1,7 +1,7 @@
 package com.wire.kalium.persistence.dao
 
 import com.wire.kalium.persistence.BaseDatabaseTest
-import com.wire.kalium.persistence.db.Database
+import com.wire.kalium.persistence.db.UserDatabaseProvider
 import com.wire.kalium.persistence.utils.stubs.newUserEntity
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.test.TestResult
@@ -19,7 +19,7 @@ class UserDAOTest : BaseDatabaseTest() {
     private val user2 = newUserEntity(id = "2")
     private val user3 = newUserEntity(id = "3")
 
-    lateinit var db: Database
+    lateinit var db: UserDatabaseProvider
 
     @BeforeTest
     fun setUp() {
