+++ persistence/src/jsMain/kotlin/com/wire/kalium/persistence/db/UserDatabaseProvider.kt
@@ -1,14 +1,14 @@
 package com.wire.kalium.persistence.db
 
-import com.wire.kalium.persistence.dao.asset.AssetDAO
 import com.wire.kalium.persistence.dao.ConversationDAO
 import com.wire.kalium.persistence.dao.MetadataDAO
 import com.wire.kalium.persistence.dao.TeamDAO
 import com.wire.kalium.persistence.dao.UserDAO
+import com.wire.kalium.persistence.dao.asset.AssetDAO
 import com.wire.kalium.persistence.dao.client.ClientDAO
 import com.wire.kalium.persistence.dao.message.MessageDAO
 
-actual class Database {
+actual class UserDatabaseProvider {
     actual val userDAO: UserDAO
         get() = TODO("Not yet implemented")
     actual val conversationDAO: ConversationDAO
